Fixed comments and change int to float.

+++ FIXED_BATCH_3/lowest_to_highest.py
@@ -1,20 +1,20 @@
-#Create empty list
+# Create empty list
 numbers_list = []
 
-#Use loop and try-except to allow users to input number and stop if the input is invalid
+# Use loop and try-except to allow users to input number and stop if the input is invalid
 while True:
     try:
-        number = int(input("Enter the number: "))
+        number = float(input("Enter the number: "))
         
-        #Update the list
+        # Update the list
         numbers_list.append(number)
 
     except ValueError:
         print("Invalid")
         break
 
-#Arrange the list from lowest to highest
+# Arrange the list from lowest to highest
 arranged_number = sorted(numbers_list)
         
-#Print the result
+# Print the result
 print(f"Result: {arranged_number}")
